Apply Dependabot dependency bumps and adapt to breaking changes
Frontend (frontend/):
- react-markdown 9.0.1 -> 10.1.0, eslint-plugin-react-refresh 0.5.2 -> 0.5.3
- react-resizable-panels ^2.1.9 -> ^4.12.2 (breaking): migrate
split-container.tsx to the v4 API (Group/Separator, orientation,
onLayoutChange with the new Layout map, percentage-string sizes).

Python tooling:
- ruff 0.15.22 -> 0.16.0: adopt the enlarged default ruleset via
behavior-preserving fixes plus justified # noqa where a fix would change
runtime behavior (DTZ naive-time scheduler, blind excepts, FastAPI
Body() default, sqlite3.Row .keys()).

+++ app/services/attribution_service.py
@@ -149,8 +149,8 @@ async def attribute_turn_cost(
             (
                 session_id,
                 project_id,
-                int(round(tokens_in * frac)),
-                int(round(tokens_out * frac)),
+                round(tokens_in * frac),
+                round(tokens_out * frac),
                 cost_usd * frac,
             ),
         )
